Formatting the classes into individual files, and placing the classes into the model directory
