+++ Private_Practice/Models/Illnesses.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Private_Practice.Models
+{
+    class Illnesses
+    {
+        // Pull Illness data from ilness API
+        string IllnessId { get; set; }
+        string IllnessName { get; set; }
+    }
+}
