+++ Private_Practice/Models/visits.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Private_Practice.Models
+{
+    class Visits
+    {
+        Guid VisitId { get; set; }
+        Guid DoctorId { get; set; }
+        Guid ClientId { get; set; }
+
+    }
+}
