+++ Private_Practice/Program.cs
@@ -31,64 +31,5 @@ namespace Private_Practice
 
 
     }
-    public enum Gender { Male, Female, Other }
-    interface IPerson
-    {
-        Guid PersonId { get; set; }
-        string Name { get; set; }
-        string Surname { get; set; }
-        Gender? Gender { get; set; }
-        int IdNumber { get; set; }
-        string Email { get; set; }
-        int ContactNo { get; set; }
-        bool Active { get; set; }
-    }
 
-    class Doctor : IPerson
-    {
-        public Guid PersonId { get; set; }
-        Guid DoctorId { get; set; }
-        string Specialist { get; set; }
-        //Set from dictionary
-        string CountryCode { get; set; }
-        public string Name { get; set; }
-        public string Surname { get; set; }
-        public Gender? Gender { get; set; }
-        public int IdNumber { get; set; }
-        public string Email { get; set; }
-        public int ContactNo { get; set; }
-        public bool Active { get; set; }
-    }
-
-
-
-    class Client : IPerson
-    {
-        public Guid ClientId { get; set; }
-        public Guid PersonId { get; set; }
-        public string IllnessId { get; set; }
-        public string Name { get; set; }
-        public string Surname { get; set; }
-        public Gender? Gender { get; set; }
-        public int IdNumber { get; set; }
-        public string Email { get; set; }
-        public int ContactNo { get; set; }
-        public bool Active { get; set; }
-
-    }
-
-    class Visits
-    {
-        Guid VisitId { get; set; }
-        Guid DoctorId { get; set; }
-        Guid ClientId { get; set; }
-
-    }
-
-    class Illness
-    {
-        // Pull Illness data from ilness API
-        string IllnessId { get; set; }
-        string IllnessName { get; set; }
-    }
 }
