+++ Private_Practice/Views/Shared/_Layout.cshtml
@@ -32,7 +32,8 @@
                 <ul class="nav navbar-nav">
                     <li><a asp-area="" asp-controller="Home" asp-action="Index">Home</a></li>
                     <li><a asp-area="" asp-controller="Home" asp-action="About">About</a></li>
-                    <li><a asp-area="" asp-controller="Home" asp-action="Contact">Contact</a></li>
+                    <li><a asp-area="" asp-controller="Doctor" asp-action="Index">Doctors</a></li>
+                    <li><a asp-area="" asp-controller="Client" asp-action="Index">Clients</a></li>
                 </ul>
             </div>
         </div>
